requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.7.0
attrs==23.2.0
Brotli==1.0.9
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cog==0.13.6
contourpy==1.3.1
cycler==0.11.0
fastapi==0.110.3
filelock==3.13.1
fonttools==4.51.0
fsspec==2024.10.0
ftfy==6.3.1
grpcio==1.62.2
h11==0.14.0
httptools==0.6.4
huggingface-hub==0.26.3
idna==3.7
importlib_metadata==8.5.0
Jinja2==3.1.4
kiwisolver==1.4.4
Markdown==3.6
MarkupSafe==2.1.3
matplotlib==3.9.2
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
open_clip_torch==2.29.0
packaging==24.1
pillow==11.0.0
pip==24.2
ply==3.11
protobuf==4.25.3
pydantic==2.10.3
pydantic_core==2.27.1
pyparsing==3.2.0
PyQt5==5.15.10
PyQt5-sip==12.13.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
sentencepiece==0.2.0
setuptools==75.1.0
sip==6.7.12
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
structlog==24.4.0
sympy==1.13.1
tensorboard==2.18.0
tensorboard_data_server==0.7.0
timm==1.0.12
tokenizers==0.21.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.67.1
transformers==4.47.0
triton==3.1.0
typing_extensions==4.12.2
unicodedata2==15.1.0
urllib3==2.2.3
uvicorn==0.32.1
uvloop==0.21.0
watchfiles==1.0.0
wcwidth==0.2.13
websockets==14.1
Werkzeug==3.1.3
wheel==0.44.0
zipp==3.21.0
